+++ beast/models/beast_vit/beast_vit_model.py
